Centralize generation emission rules
